Adds `pyatlan-formatter`
I observed that the pre-commit script only executes on Git-tracked files.
However, this poses an issue when we generate new modules and attempt to run
the formatter on them, as it requires tracking (staging) those files before running
`pre-commit run --all-files`. This adds unnecessary steps for users who may not be familiar with pre-commit.

To address this, I propose the addition of a simple shell script,
`./pyatlan-formatter`, which can handle formatting tasks for both tracked and untracked .